service/collection/tls/requirements.txt
grpcio
grpcio-tools
grpcio-reflection
pyyaml
pytest
coverage
oauthlib
requests_oauthlib
